Revert "Merge remote-tracking branch 'naddiseo/master'"
This reverts commit 22ef6fc7f84869d3f2efb646ba080e29ae211d30, reversing
changes made to f7cba9bb737975ea9260cab427c13684ffd233b2.

+++ include/math/CRC.h
@@ -19,12 +19,10 @@
 #ifndef HPL_CRC_H
 #define HPL_CRC_H
 
-#include <stdint.h>
-
 namespace hpl {
 	//----------------------------------------
 
-	typedef uint32_t tCRCKey;
+	typedef unsigned long tCRCKey;
 
 	//----------------------------------------
 
@@ -32,9 +30,9 @@ namespace hpl {
 	{
 	public:
 		cCRCTable () : mKey (0) {}
-
+		
 		void Init (tCRCKey key);
-
+		
 		tCRCKey operator [] (unsigned i){return mTable [i];}
 
 	private:
@@ -51,7 +49,7 @@ namespace hpl {
 		{
 			mTable.Init (key);
 		}
-
+		
 		void PutByte (unsigned aByte);
 
 		tCRCKey Done ()
@@ -63,7 +61,7 @@ namespace hpl {
 
 	private:
 		static cCRCTable mTable;
-		tCRCKey mKey;
+		tCRCKey mKey;	
 		tCRCKey mRegister;
 	};
